Switch JSON Canonicalization Library
For JSON canonicalization, which is needed for identifier generation of
objects with extensions, the library github.com/ugorji/go/codec was
initially used. This library include more than just canonicalization
functionality. So to reduce dependent code, a decision to switch to a
smaller library was made. The library picked was
cyberphone/json-canonicalization. This library has implementations of
the RFCs in different languages.